Rename files to kebab-case convention
